Edit ViewModel

+++ src/UniAlumni.Business/Services/EventRegistrationService/EventRegistrationSvc.cs
@@ -1,8 +1,10 @@
 ﻿using System;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using UniAlumni.DataTier.Common.Enum;
+using UniAlumni.DataTier.Common.Exception;
 using UniAlumni.DataTier.Models;
 using UniAlumni.DataTier.Repositories.AlumniRepo;
 using UniAlumni.DataTier.Repositories.EventRegistrationRepo;
@@ -34,23 +36,30 @@ namespace UniAlumni.Business.Services.EventRegistrationService
                 Alumnus alumnus = await queryAlumni.FirstOrDefaultAsync();
                 if (alumnus == null || alumnus.Status != (byte?) AlumniEnum.AlumniStatus.Active)
                 {
-                    throw new Exception("Alumni not exist or not active");
+                    throw new MyHttpException(StatusCodes.Status404NotFound,"Alumni not exist or not active");
                 }
 
                 IQueryable<Event> queryEvent = _eventRepository.Table.Where(e => e.Id == eventId);
                 Event eventDetail = await queryEvent.FirstOrDefaultAsync();
-                if (eventDetail == null || eventDetail.Status != (byte?) EventEnum.EventStatus.Delete)
+                if (eventDetail == null || eventDetail.Status != (byte?) EventEnum.EventStatus.RegistrationStart)
                 {
-                    throw new Exception("EventNotFound");
+                    throw new MyHttpException(StatusCodes.Status404NotFound,"Event not exist or not start register");
                 }
 
                 EventRegistration newEventRegistration = new EventRegistration() {
                     AlumniId = alumniId, 
                     EventId = eventId,
-                    Status = (byte?) EventRegistrationEnum.EventRegistrationStatus.Pending};
+                    Status = (byte?) EventRegistrationEnum.EventRegistrationStatus.Joined,
+                    RegisteredDate = DateTime.Now
+                };
 
                 await _eventRegistrationRepository.InsertAsync(newEventRegistration);
                 await _eventRegistrationRepository.SaveChangesAsync();
+            }
+            else
+            {
+                throw new MyHttpException(StatusCodes.Status204NoContent,"Event has been registed");
+
             }
             
         }
@@ -60,6 +69,10 @@ namespace UniAlumni.Business.Services.EventRegistrationService
             IQueryable<EventRegistration> queryEventRegistration = _eventRegistrationRepository.Table.Where(er
                 => er.AlumniId == alumniId && er.EventId == eventId);
             EventRegistration eventRegistration = await queryEventRegistration.FirstOrDefaultAsync();
+            if (eventRegistration == null)
+            {
+                throw new MyHttpException(StatusCodes.Status404NotFound, "You cannot leave");
+            }
             _eventRegistrationRepository.Delete(eventRegistration);
             await _eventRegistrationRepository.SaveChangesAsync();
         }
